Release 4.5.0

+++ services/bidding_service/base_generate_bids_reactor.h
@@ -24,6 +24,7 @@
 #include "services/bidding_service/data/runtime_config.h"
 #include "services/common/clients/code_dispatcher/request_context.h"
 #include "services/common/code_dispatch/code_dispatch_reactor.h"
+#include "services/common/constants/common_constants.h"
 #include "services/common/loggers/request_log_context.h"
 #include "services/common/util/request_response_constants.h"
 
@@ -31,7 +32,7 @@ namespace privacy_sandbox::bidding_auction_servers {
 
 inline constexpr char kDispatchHandlerFunctionNameWithCodeWrapper[] =
     "generateBidEntryFunction";
-inline constexpr int kBytesMultiplyer = 1024;
+inline constexpr long kBytesMultiplyer = 1024;
 
 // Returns up the index of the provided enum as int from the underlying enum
 // storage.
